requirements.txt
scikit-learn
pandas
numpy
PyYAML
streamlit
-e .


